Added basebranch git config parameter
All of the mock api / stub_json parameter had to be adjusted to
reflect the new `base: ref: key` response. This is because the
tests run in tmpdirs which set the git config parameters each time
using the common in.rb/output function, and if we don't set them
for each of the tests this function will fail.