clone base branch instead of remote default branch
Clone the base branch instead of the default remote branch. This is
mostly useful when the base branch is not the default remote branch,
especially when the depth filter is used; it ensures that the base
branch's ref is available so that it can be diffed with the branch ref.