Prevent `check` from missing versions
There are situations where new versions of the PR resource will get
lost. This happens because the versions that are returned are sorted by
when they were last updated. Comments on PRs can affect the update time.
This means that between checks, if the version Concourse passes to the
`check` script is commented on, Concourse will only consider anything
updated after this PR a valid version. If PRs were updated between the
time of the last check and the time of this update, they will be lost.
This change puts the version Concourse passes to the `check` script at
the start of the list of versions it returns. This ensures that a valid
version does not get skipped for this reason.