Implement pulling pull requests which are created from a specific head
This is the other way around from the base branch option. It will allow to
trigger for pull requests only, which are created from a specific branch.
This can, e.g., used to trigger a specific branch pipeline in concourse for
pull requests, too, without triggering for all other head branches.

Fixes #142